ipv6-range: For a prefixlen < 64 display the number of /64's
When given a network with a prefix length of < 64, display the number of
/64 networks that provides. e.g a /48 provides 65536 /64's.

This is perhaps useful as the recommended network block size to be
assigned to end users is a /48 from which they can allocate 65536 /64's
each containing 18,446,744,073,709,551,616 addresses. With the last 64
bits being the recommended size of the host part. Some ISP's may only
give a /56 rather than a /48.

Example output:

    $ ./ipv6-range 2001:db8:1::/48
    Network : 2001:db8:1::/48 (65536 /64 networks)
    Start   : 2001:0db8:0001:0000:0000:0000:0000:0000
    End     : 2001:0db8:0001:ffff:ffff:ffff:ffff:ffff

    $ ./ipv6-range 2001:db8:1::/56
    Network : 2001:db8:1::/56 (256 /64 networks)
    Start   : 2001:0db8:0001:0000:0000:0000:0000:0000
    End     : 2001:0db8:0001:00ff:ffff:ffff:ffff:ffff

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>