mac-to-eui64: Generate a modified EUI-64 from a MAC address
This is a simple tool that takes a MAC address and generates the
associated modified EUI-64 identifier.

E.g, given the MAC address 00:21:2f:b5:6e:10 it will output
02:21:2f:ff:fe:b5:6e:10

It takes the MAC address, inserts ff:fe into the middle and flips the
universal/local (U/L) bit, the seventh bit from the left.

This is defined in RFC 4291[0] Appendix A.

These are used when auto generating IPv6 addresses, e.g link local and
SLAAC.

[0] - https://tools.ietf.org/html/rfc4291

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>