Makefile: Add a test for the glibc version
The gen-ula program uses clock_gettime(2), under glibc < 2.17 this
requires linking against -lrt.

Add a test in the Makefile to determine the glibc version in use. If
it's < 2.17 then set the LIBS variable to -lrt otherwise leave it blank
and add it to the gen-ula compile command.

This fix's building on CentOS 6 for example.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>