ipv6-isin: Re-work it
This tool given a network/prefixlen and an address will check whether
the address falls within the given network.

However this was showing some problems with smaller prefix's and
prefix's not a multiple of 8. e.g with the previous version

    $ ./ipv6-isin 2001:db8:beef::/60 2001:db8:beef:10::cafe
    Yes
    $ ./ipv6-isin fec0::/10 fec0::a6d3:6545:8d49:8a99
    No

In the first case the valid address range is

    2001:0db8:beef:0000:0000:0000:0000:0000 -
    2001:0db8:beef:000f:ffff:ffff:ffff:ffff

in the second instance the valid address range is

   fec0:0000:0000:0000:0000:0000:0000:0000 -
   feff:ffff:ffff:ffff:ffff:ffff:ffff:ffff

With the fixed version we now get

    $ ./ipv6-isin 2001:db8:beef::/60 2001:db8:beef:10::cafe
    No
    $ ./ipv6-isin fec0::/10 fec0::a6d3:6545:8d49:8a99
    Yes

This commit replaces the tcp wrappers based code with code based on the
ngx_ptocidr() & ngx_http_access_inet6() functions from nginx.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>