Tag the license and README files as linguist-documentation
In GitHub, ipv6-tools is showing up as mostly Groff!

It seems linguist (what GitHub use to determine what languages are used
within projects) thinks the LGPLv2.1 file is groff! No idea why this
is, file says

    LGPLv2.1: ASCII text

so nothing untoward there. Maybe because it's named with a '.1' who
knows...

Anyway slightly irritating, so lets add some .gitattributes entries to
mark the two license file and the readme for good measure as
linguist-documentation which should tell linguist to ignore these files.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>