Add and switch to "short types"
Add a cut down version of the short_types.h[0] header file from the CCAN
project.

This adds the following types as used by the likes of the Linux kernel

    u64, s64, u32, s32, u16, s16, u8 & s8

They are typdef'd to their C99 fixed width equivalents, u being unsigned
and s signed and the number being the size in bits. They are shorter to
type and perhaps easier on the eyes then the likes of uint64_t.

For VIM users, you can add the following to your .vim/after/syntax/c.vim
to get syntax highlighting for these new types

    syn keyword cType u64
    syn keyword cType s64
    syn keyword cType u32
    syn keyword cType s32
    syn keyword cType u16
    syn keyword cType s16
    syn keyword cType  u8
    syn keyword cType  s8

[0] - https://ccodearchive.net/info/short_types.html

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>