ipv6-range: Display the host part in bold
If you're not familiar with this stuff sometimes it won't be clear just
which bit is the network and which is the host part.

So to make it more readily clear which bits are which we print the host
part (the bit the user can change) in bold. e.g

    $ ./ipv6-range 2001:db8::/112
    Network : 2001:db8::/112
    Start   : 2001:0db8:0000:0000:0000:0000:0000:0000
    End     : 2001:0db8:0000:0000:0000:0000:0000:ffff

it's not immediately clear which bits the user can set for the host.
However now the host part is actually displayed in bold on the console
so you would see that last group (:0000 - :ffff) displayed as bold.

To do that we needed to change the loops to print out at the nibble
level to track when we cross nibble boundaries and things like

    $ ./ipv6-range 2001:db8::/61
    Network : 2001:db8::/61 (8 /64 networks)
    Start   : 2001:0db8:0000:0000:0000:0000:0000:0000
    End     : 2001:0db8:0000:0007:ffff:ffff:ffff:ffff

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>