ipv6-range: Show number of /48 and /60 networks
Seeing as a /48 is what is recommended ISPs hand out to users, it's a
useful thing to show for prefixes < 48.

Also show the number of /60s for no other reason than my ISP lets me
allocate both /64s and /60s from my /48.

I've also made it work for a prefix of /0, not useful in practice but
for completeness.

The output has also changed slightly to accommodate the longer network
string.

Old output

    $ ./ipv6-range 2001:db8::/32
    Network : 2001:db8::/32 (16777216 /56 networks, 4294967296 /64 networks)
    Start   : 2001:0db8:0000:0000:0000:0000:0000:0000
    End     : 2001:0db8:ffff:ffff:ffff:ffff:ffff:ffff

    A /64 provides 18,446,744,073,709,551,616 addresses

new output

    $ ./ipv6-range 2001:db8::/32
    Network : 2001:db8::/32
              (65536 /48s, 16777216 /56s, 268435456 /60s, 4294967296 /64s)
    Start   : 2001:0db8:0000:0000:0000:0000:0000:0000
    End     : 2001:0db8:ffff:ffff:ffff:ffff:ffff:ffff

    A /64 provides 18,446,744,073,709,551,616 addresses

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>