mac-to-eui64: Fix the displaying of some IPv6 addresses
Given a MAC address of de:ad:be:ef:00:01 we would print

    (SLAAC Host)    : ::dcad:beff:feef:01
    (link-local)    : fe80::dcad:beff:feef:01

While not wrong, the :01 would be better shown as just :1

However, given 00:10:a4:01:23:45 we would print

    (SLAAC Host)    : ::210:a4ff:fe1:2345
    (link-local)    : fe80::210:a4ff:fe1:2345

this _is_ wrong. The :fe1: should be :fe01:

The simplest, and perhaps correct way in any case, to fix this is to
simply display the addresses in uncompressed form (the other tools in
here tend to display in uncompressed form and there is always the
ipv6-fmt program) by ensuring we 0 pad each component of the EUI-64
address when combining the pairs.

So now the above two examples become

    (SLAAC Host)    : ::dcad:beff:feef:0001
    (link-local)    : fe80::dcad:beff:feef:0001

    (SLAAC Host)    : ::0210:a4ff:fe01:2345
    (link-local)    : fe80::0210:a4ff:fe01:2345

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>