ipv6-gen-slaac: Handle < /64 IPv6 network addresses
When given an address like, 2001:db8::, we would print

    SLAAC : d00::100:0:0:0

this was due to trying to handle addresses like 2001:db8:a;b::, where
we'd chop the trailing ';' off.

That of course breaks for networks less than a /64, e.g  2001:db8::,
where we need to leave the '::'.

We now handle both those cases.

It may also be useful to handle networks specified with a prefixlen, e.g
2001:db8::/32, so we now handle that case as well.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>